// src/types/qiita-types.d.ts
export interface User {
  /**
   * self-description
   */
  description: null | string;
  /**
   * Facebook ID
   */
  facebook_id: null | string;
  /**
   * Followees count
   */
  followees_count: number;
  /**
   * Followers count
   */
  followers_count: number;
  /**
   * GitHub ID
   */
  github_login_name: null | string;
  /**
   * User ID
   */
  id: string;
  /**
   * How many items a user posted on qiita.com (Items on Qiita Team are not included)
   */
  items_count: number;
  /**
   * LinkedIn ID
   */
  linkedin_id: null | string;
  /**
   * Location
   */
  location: null | string;
  /**
   * Customized user name
   */
  name: null | string;
  /**
   * Organization which a user belongs to
   */
  organization: null | string;
  /**
   * Unique integer ID
   */
  permanent_id: number;
  /**
   * Profile image URL
   */
  profile_image_url: string;
  /**
   * A flag whether this user is configured as team-only
   */
  team_only: boolean;
  /**
   * Twitter screen name
   */
  twitter_screen_name: null | string;
  /**
   * Website URL
   */
  website_url: null | string;

  [k: string]: unknown;
}
// ...
export interface Group {
  /**
   * Date-time when this data was created
   */
  created_at: string;
  /**
   * A group unique ID
   */
  id: number;
  /**
   * Group name for display.
   */
  name: string;
  /**
   * A flag to tell which this group is private or public.
   */
  private: boolean;
  /**
   * Date-time when this data was updated
   */
  updated_at: string;
  /**
   * Unique name on a team.
   */
  url_name: string;

  [k: string]: unknown;
}

/**
 * Represents an item posted from a user
 */
export interface Item {
  /**
   * Item body in HTML
   */
  rendered_body: string;
  /**
   * Item body in Markdown
   */
  body: string;
  /**
   * A flag whether this item is co-edit mode (only available on Qiita Team)
   */
  coediting: boolean;
  /**
   * Comments count
   */
  comments_count: number;
  /**
   * Date-time when this data was created
   */
  created_at: string;
  group: Group;
  /**
   * An unique item ID
   */
  id: string;
  /**
   * Likes count (only available on Qiita)
   */
  likes_count: number;
  /**
   * A flag whether this item is private (only available on Qiita)
   */
  private: boolean;
  /**
   * Emoji reactions count (only availabble on Qiita Team)
   */
  reactions_count: number;
  /**
   * A list of tags
   */
  tags: {
    name: string;
    versions?: string[];
    [k: string]: unknown;
  }[];
  /**
   * The title of this item
   */
  title: string;
  /**
   * Date-time when this data was updated
   */
  updated_at: string;
  /**
   * The URL of this item
   */
  url: string;
  user: User;
  /**
   * The number of views.
   */
  page_views_count: null | number;

  [k: string]: unknown;
}